requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pandas==2.0.3
scikit-learn==1.3.0
python-docx==0.8.11
openai==1.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
eventlet==0.33.3
